pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "gcloud-oauth2-cli"
package-mode = false
version = "0.1.0"
description = "An example demonstrating using OAuth2.0 to authenticate to Google Cloud."
authors = ["Xinyu Ma <ma.xinyu.26a@kyoto-u.jp>"]
license = "Apache-2.0"
readme = ["README.md"]

[tool.poetry.dependencies]
python = "^3.12"
cryptography = "^40.0.2"
aiohttp = "^3.8.5"
pyjwt = "^2.7.0"
